Moved BrandDerivable
