+++ lib/design_system/generic.rb
@@ -2,7 +2,6 @@
 
 # This is the generic (unbranded) adapter for design system
 
-require_relative 'generic/builders/concerns/brand_derivable'
 require_relative 'generic/builders/button'
 require_relative 'generic/builders/callout'
 require_relative 'generic/builders/code'
