+++ lib/design_system/generic/builders/base.rb
@@ -5,7 +5,7 @@ module DesignSystem
     module Builders
       # This is the base class for design system builders.
       class Base
-        include DesignSystem::Generic::Builders::Concerns::BrandDerivable
+        include DesignSystem::BrandDerivable
         include DesignSystem::Helpers::CssHelper
 
         delegate :button_tag, :capture, :content_for, :content_tag, :link_to, :link_to_unless_current, to: :@context
