--- lib/design_system/generic/builders/concerns/brand_derivable.rb
@@ -1,14 +1,0 @@
-module DesignSystem
-  module Generic
-    module Builders
-      module Concerns
-        # Mixin module to make brand method available to builder classes
-        module BrandDerivable
-          def brand
-            self.class.name.split('::')[1].underscore
-          end
-        end
-      end
-    end
-  end
-end
